refactor: move repl info code to separate module
